yohane-cli/pyproject.toml
[project]
name = "yohane-cli"
dynamic = ["version"]
description = "Typer CLI for yohane"
readme = "../README.md"
license = "MIT"
authors = [{ name = "NextFire", email = "git@yuru.moe" }]
requires-python = ">=3.10"
dependencies = ["typer-slim[standard]==0.12.5", "yt-dlp[default]==2024.10.22"]

[project.urls]
homepage = "https://github.com/Japan7/yohane"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"
raw-options = { root = ".." }
